Corrected Urgency to not use magic
